First upload
The basic RAG answer generation function is implemented, but the function of processing loaded emails still needs to be optimized. loadfile.py reads the relevant regulations and stores the content in Knowlege Base.txt. gmailload.py can read the selected email. raganswer-static.py can read the questions in questions.txt and generate answers. raganswer-dynamic.py is used for continuous questioning, and the actual work on the server when used simulates real continuous communication and answering